fix(gemini): split image request helper and complete provenance
